feat(scraping): Phase 1 - Spotify/YouTube/SoundCloud scrapers + /scrape + UI
Scrapers (verified live):
- spotify.py: monthly listeners via Playwright render (token endpoint is TOTP-signed)
- youtube.py: subscribers + top video via Data API v3
- soundcloud.py: followers + top track via api-v2 (client_id auto-extracted)

Orchestration & API:
- scrape_service.py: per-artist dispatch, 24h TTL cache, partial results, DB persist
- POST /scrape (data-api) + Next.js /api/scrape proxy
- base.py: env-tunable throttle, render_html wait_text

Schema/infra:
- count columns widened to BIGINT (YouTube views exceed INTEGER's 2.1B cap)
- docker-compose: pass YOUTUBE_API_KEY, drop obsolete version key
- artist detail page: 'Sources & Scraping' panel + cross-platform metrics

+++ data-api/schema.sql
@@ -7,24 +7,25 @@ CREATE TABLE artists (
     genres TEXT NULL,
     images JSON NULL,
     external_urls JSON NULL,
-    monthly_listeners INTEGER NULL,
+    monthly_listeners BIGINT NULL,
     spotify_id VARCHAR(255) NULL,
     youtube_id VARCHAR(255) NULL,
     top_track_id VARCHAR(255) NULL,
     top_track_name VARCHAR(255) NULL,
     top_track_popularity INTEGER NULL,
 
-    -- Cross-platform scraping metrics (see ai/SCRAPING_PLAN.md)
-    youtube_subscribers INTEGER NULL,
+    -- Cross-platform scraping metrics (see ai/SCRAPING_PLAN.md).
+    -- Counts are BIGINT: views/plays/likes routinely exceed INTEGER's 2.1B cap.
+    youtube_subscribers BIGINT NULL,
     youtube_top_video_title VARCHAR(255) NULL,
-    youtube_top_video_views INTEGER NULL,
-    soundcloud_followers INTEGER NULL,
+    youtube_top_video_views BIGINT NULL,
+    soundcloud_followers BIGINT NULL,
     soundcloud_top_track VARCHAR(255) NULL,
-    soundcloud_top_track_plays INTEGER NULL,
-    instagram_followers INTEGER NULL,
-    tiktok_followers INTEGER NULL,
-    tiktok_likes INTEGER NULL,
-    x_followers INTEGER NULL,
+    soundcloud_top_track_plays BIGINT NULL,
+    instagram_followers BIGINT NULL,
+    tiktok_followers BIGINT NULL,
+    tiktok_likes BIGINT NULL,
+    x_followers BIGINT NULL,
     social_links JSONB NULL,
     scrape_meta JSONB NULL
 );
